dev logging fix, moralis stream all networks support, telegram warning, wallet validation

+++ crypto-tracker/database/migrations/2026_02_17_000001_create_networks_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug')->unique();
-            $table->integer('chain_id')->nullable();
+            $table->bigInteger('chain_id')->nullable();
             $table->string('currency_symbol');
             $table->string('explorer_url')->nullable();
             $table->boolean('is_active')->default(true);
